Bump version to 0.0.6
Changed behavior of from_now option in param(). Previously, this caused
setValueAtTime to be called using the param.value and
context.currentTime. The problem was that param.value did not reflect
changes to the param that had been scheduled on the same tick. So for
example, using param( ‘gain’, 1) followed by param({ gain:0, ramp:
‘linear’, from_now: true, at: 1 }) would result in no change if the
value of param gain was already at zero. This is fixed by deferring the
calls to setValueAtTime and the ramp functions by 4ms via setTimeout.